requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
black==24.8.0
click==8.1.7
decorator==5.1.1
Django==5.1.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
executing==2.1.0
inflection==0.5.1
ipython==8.29.0
isort==5.13.2
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tomli==2.0.2
traitlets==5.14.3
typing_extensions==4.12.2
drf-spectacular~=0.27.2
django-celery-results==2.5.1
uritemplate==4.1.1
wcwidth==0.2.13
